Moved the projectlaunch files to a nested directory
This should prevent conflicts with similarly named files in other plugins.
